Expose managed versions for spring libraries
We currently have versions for splring and spring-boot defined in the core catalog

This becomes a chore to keep in line when renovate spots a new release of Spring.

This change exports these spring libraries as managed versions in the BOM, so we
can remove these versions in core when we integrate the micronaut-spring BOM into
core.

Currently, the bom for this module is unused in core.

I believe there was also an issue whereby this app had upgraded to 5.3.20 of Spring,
but the property in gradle.properties was still on 5.3.9.  I have removed this property
and the resolution strategy, and added the new method required by MicronautApplicationContext.