Add support HttpExchange annotations (#652)
It's not full support of these annotations, because they primarily describe the client's behavior, i.e. describe its requests for specific endpoints. But Spring allows using these annotations to describe a single API interface. Therefore, it is important for us to have correct mapping of these annotations specifically for the server, i.e. when they are used in the controller to describe endpoints.